Updated visualization tools. We now use a Visdom Server to visualize all the data (that are, however, exposed into the terminal to have a fast check). Moreover, to see well the result, we print the confusion matrix for the cumulative test (from which the others can be restored)
I also added a order file for OfficeHome (1 sequential + 9 random orders)

Todo:
Find a better way to save results that prevent opening n_runs files for each method tested.

+++ data/idadataloader.py
@@ -61,7 +61,7 @@ class IDADataloader(AbstractIncrementalDataloader):
         if order_file is None:
             self.full_order = [np.arange(self.num_classes)]  # if not specified go from zero to num_classes in order
         else:
-            self.full_order = np.load(order_file)
+            self.full_order = np.load(order_file).astype(int)
 
         # init parameters
         self.iteration = 0
